Add: New logo Add: Now has tags in drinks Fix: Now fragments scroll as expected Fix: Now does not crash on valid server info Fix: Now sorts drinks and produts by asc

+++ app/src/main/java/com/rannasta_suomeen/data_classes/Cabinet.kt
@@ -56,7 +56,7 @@ class Infinite: OwnedAmount(){
 class Some(val x: Int): OwnedAmount(){
     override fun show(settings: Settings): String {
         Log.d("Owned", "Amount_ml is: $x")
-        return UnitType.ml.displayInDesiredUnit(x,settings.prefUnit)
+        return UnitType.ml.displayInDesiredUnit(x.toDouble(),settings.prefUnit)
     }
 }
